Move requirements and request inputs
