+++ study/api-src/org/labkey/api/specimen/SpecimenMigrationService.java
@@ -24,32 +24,33 @@ public interface SpecimenMigrationService
     ActionURL getAutoReportListURL(Container c);
     ActionURL getConfigureRequestabilityRulesURL(Container c);
     ActionURL getManageActorsURL(Container c);
+    ActionURL getManageDefaultReqsSettingsURL(Container c);
     ActionURL getManageDisplaySettings(Container c);
     ActionURL getManageNotificationsURL(Container c);
     ActionURL getManageRepositorySettingsURL(Container c);
-    ActionURL getManageDefaultReqsSettingsURL(Container c);
     ActionURL getManageRequestStatusURL(Container c, int requestId);
+    ActionURL getManageRequestURL(Container c, int requestId, @Nullable ActionURL returnUrl);
     ActionURL getManageStatusesURL(Container c);
     ActionURL getOverviewURL(Container c);
     ActionURL getShowGroupMembersURL(Container c, int rowId, @Nullable Integer locationId, @Nullable ActionURL returnUrl);
     ActionURL getShowSearchURL(Container c, boolean showVials);
-    ActionURL getSpecimensURL(Container c);
     ActionURL getSpecimenEventsURL(Container c, ActionURL returnUrl);
     ActionURL getSpecimenRequestConfigRequiredURL(Container c);
     ActionURL getSpecimenRequestEventDownloadURL(SpecimenRequestEvent event, String name);
+    ActionURL getSpecimensURL(Container c);
     ActionURL getUploadSpecimensURL(Container c);
     ActionURL getViewRequestsURL(Container c);
-    ActionURL getManageRequestURL(Container c, int requestId, @Nullable ActionURL returnUrl);
 
-    Class<? extends Controller> getShowCreateSpecimenRequestActionClass();
+    Class<? extends Controller> getClearCommentsActionClass();
+    Class<? extends Controller> getDeleteRequestActionClass();
     Class<? extends Controller> getExtendedSpecimenRequestActionClass();
-    Class<? extends Controller> getRemoveRequestSpecimensActionClass();
     Class<? extends Controller> getImportVialIdsActionClass();
     Class<? extends Controller> getManageRequestActionClass();
-    Class<? extends Controller> getClearCommentsActionClass();
-    Class<? extends Controller> getUpdateCommentsActionClass();
-    Class<? extends Controller> getDeleteRequestActionClass();
-    Class<? extends Controller> getSubmitRequestActionClass();
-    Class<? extends Controller> getRequestHistoryActionClass();
     Class<? extends Controller> getManageRequestInputsActionClass();
+    Class<? extends Controller> getManageRequirementActionClass();
+    Class<? extends Controller> getRemoveRequestSpecimensActionClass();
+    Class<? extends Controller> getRequestHistoryActionClass();
+    Class<? extends Controller> getShowCreateSpecimenRequestActionClass();
+    Class<? extends Controller> getSubmitRequestActionClass();
+    Class<? extends Controller> getUpdateCommentsActionClass();
 }
