+++ study/src/org/labkey/study/controllers/specimen/SpecimenController.java
@@ -47,7 +47,6 @@ import org.labkey.api.specimen.RequestEventType;
 import org.labkey.api.specimen.SpecimenManagerNew;
 import org.labkey.api.specimen.SpecimenMigrationService;
 import org.labkey.api.specimen.SpecimenRequestManager;
-import org.labkey.api.specimen.SpecimenRequestManager.SpecimenRequestInput;
 import org.labkey.api.specimen.Vial;
 import org.labkey.api.specimen.actions.IdForm;
 import org.labkey.api.specimen.actions.ParticipantCommentForm;
@@ -59,16 +58,12 @@ import org.labkey.api.specimen.model.SpecimenRequestActor;
 import org.labkey.api.specimen.model.SpecimenRequestEvent;
 import org.labkey.api.specimen.notifications.ActorNotificationRecipientSet;
 import org.labkey.api.specimen.notifications.DefaultRequestNotification;
-import org.labkey.api.specimen.notifications.NotificationRecipientSet;
 import org.labkey.api.specimen.query.SpecimenQueryView;
 import org.labkey.api.specimen.requirements.SpecimenRequest;
 import org.labkey.api.specimen.requirements.SpecimenRequestRequirement;
 import org.labkey.api.specimen.requirements.SpecimenRequestRequirementProvider;
 import org.labkey.api.specimen.security.permissions.EditSpecimenDataPermission;
-import org.labkey.api.specimen.security.permissions.ManageNewRequestFormPermission;
 import org.labkey.api.specimen.security.permissions.ManageRequestsPermission;
-import org.labkey.api.specimen.settings.RequestNotificationSettings;
-import org.labkey.api.specimen.settings.SettingsManager;
 import org.labkey.api.study.CohortFilter;
 import org.labkey.api.study.SpecimenUrls;
 import org.labkey.api.study.Study;
@@ -83,7 +78,6 @@ import org.labkey.api.util.DateUtil;
 import org.labkey.api.util.Pair;
 import org.labkey.api.view.ActionURL;
 import org.labkey.api.view.DataView;
-import org.labkey.api.view.GridView;
 import org.labkey.api.view.HtmlView;
 import org.labkey.api.view.HttpView;
 import org.labkey.api.view.InsertView;
@@ -94,7 +88,6 @@ import org.labkey.api.view.RedirectException;
 import org.labkey.api.view.UnauthorizedException;
 import org.labkey.api.view.UpdateView;
 import org.labkey.api.view.VBox;
-import org.labkey.api.view.ViewContext;
 import org.labkey.study.controllers.BaseStudyController;
 import org.labkey.study.model.DatasetDefinition;
 import org.labkey.study.model.SecurityType;
@@ -413,98 +406,6 @@ public class SpecimenController extends BaseStudyController
         return requestedVials;
     }
 
-    public static class RequirementForm extends IdForm
-    {
-        private int _requirementId;
-
-        public int getRequirementId()
-        {
-            return _requirementId;
-        }
-
-        public void setRequirementId(int requirementId)
-        {
-            _requirementId = requirementId;
-        }
-    }
-
-    public static class ManageRequirementForm extends RequirementForm
-    {
-        private boolean _complete;
-        private String _comment;
-        private String[] _notificationIdPairs;
-        private boolean _emailInactiveUsers;
-
-        public String getComment()
-        {
-            return _comment;
-        }
-
-        public void setComment(String comment)
-        {
-            _comment = comment;
-        }
-
-        public boolean isComplete()
-        {
-            return _complete;
-        }
-
-        public void setComplete(boolean complete)
-        {
-            _complete = complete;
-        }
-
-        public String[] getNotificationIdPairs()
-        {
-            return _notificationIdPairs;
-        }
-
-        public void setNotificationIdPairs(String[] notificationIdPairs)
-        {
-            _notificationIdPairs = notificationIdPairs;
-        }
-
-        public boolean isEmailInactiveUsers()
-        {
-            return _emailInactiveUsers;
-        }
-
-        public void setEmailInactiveUsers(boolean emailInactiveUsers)
-        {
-            _emailInactiveUsers = emailInactiveUsers;
-        }
-    }
-
-    @RequiresPermission(ManageRequestsPermission.class)
-    public class DeleteRequirementAction extends FormHandlerAction<RequirementForm>
-    {
-        @Override
-        public void validateCommand(RequirementForm target, Errors errors)
-        {
-        }
-
-        @Override
-        public boolean handlePost(RequirementForm form, BindException errors) throws Exception
-        {
-            SpecimenRequestRequirement requirement =
-                    SpecimenRequestRequirementProvider.get().getRequirement(getContainer(), form.getRequirementId());
-            if (requirement.getRequestId() == form.getId())
-            {
-                SpecimenRequestManager.get().deleteRequestRequirement(getUser(), requirement);
-                return true;
-            }
-
-            return false;
-        }
-
-        @Override
-        public ActionURL getSuccessURL(RequirementForm requirementForm)
-        {
-            return getManageRequestURL(requirementForm.getId());
-        }
-    }
-
     @RequiresPermission(ManageRequestsPermission.class)
     public class DeleteMissingRequestSpecimensAction extends FormHandlerAction<IdForm>
     {
@@ -528,168 +429,6 @@ public class SpecimenController extends BaseStudyController
         public ActionURL getSuccessURL(IdForm requirementForm)
         {
             return getManageRequestURL(requirementForm.getId());
-        }
-    }
-
-    public class ManageRequirementBean
-    {
-        private final GridView _historyView;
-        private final SpecimenRequestRequirement _requirement;
-        private final boolean _requestManager;
-        private final List<ActorNotificationRecipientSet> _possibleNotifications;
-        private final boolean _finalState;
-
-        public ManageRequirementBean(ViewContext context, SpecimenRequest request, SpecimenRequestRequirement requirement)
-        {
-            _requirement = requirement;
-            _possibleNotifications = getUtils().getPossibleNotifications(request);
-            SimpleFilter filter = new SimpleFilter(FieldKey.fromParts("RequestId"), requirement.getRequestId());
-            filter.addCondition(FieldKey.fromParts("RequirementId"), requirement.getRowId());
-            _requestManager = context.getContainer().hasPermission(context.getUser(), ManageRequestsPermission.class);
-            _historyView = getUtils().getRequestEventGridView(context.getRequest(), null, filter);
-            _finalState = SpecimenRequestManager.get().isInFinalState(request);
-        }
-        
-        public boolean isDefaultNotification(ActorNotificationRecipientSet notification)
-        {
-            RequestNotificationSettings settings = SettingsManager.get().getRequestNotificationSettings(getContainer());
-            if (settings.getDefaultEmailNotifyEnum() == RequestNotificationSettings.DefaultEmailNotifyEnum.All)
-                return true;        // All should be checked
-            else if (settings.getDefaultEmailNotifyEnum() == RequestNotificationSettings.DefaultEmailNotifyEnum.None)
-                return false;       // None should be checked
-            // Otherwise use Actor Notification
-
-            Integer requirementActorId = _requirement.getActorId();
-            Integer notificationActorId = notification.getActor() != null ? notification.getActor().getRowId() : null;
-            Integer requirementSiteId = _requirement.getSiteId();
-            Integer notificationSiteId = notification.getLocation() != null ? notification.getLocation().getRowId() : null;
-            return nullSafeEqual(requirementActorId, notificationActorId) &&
-                    nullSafeEqual(requirementSiteId, notificationSiteId);
-        }
-
-        public GridView getHistoryView()
-        {
-            return _historyView;
-        }
-
-        public SpecimenRequestRequirement getRequirement()
-        {
-            return _requirement;
-        }
-
-        public List<ActorNotificationRecipientSet> getPossibleNotifications()
-        {
-            return _possibleNotifications;
-        }
-
-        public boolean isRequestManager()
-        {
-            return _requestManager;
-        }
-
-        public boolean isFinalState()
-        {
-            return _finalState;
-        }
-    }
-
-    @RequiresPermission(ReadPermission.class)
-    public class ManageRequirementAction extends FormViewAction<ManageRequirementForm>
-    {
-        private SpecimenRequest _specimenRequest;
-
-        @Override
-        public void validateCommand(ManageRequirementForm target, Errors errors)
-        {
-        }
-
-        @Override
-        public ModelAndView getView(ManageRequirementForm form, boolean reshow, BindException errors)
-        {
-            _specimenRequest = SpecimenRequestManager.get().getRequest(getContainer(), form.getId());
-            final SpecimenRequestRequirement requirement =
-                    SpecimenRequestRequirementProvider.get().getRequirement(getContainer(), form.getRequirementId());
-            if (_specimenRequest == null || requirement == null || requirement.getRequestId() != form.getId())
-                throw new NotFoundException();
-
-            return new JspView<>("/org/labkey/study/view/specimen/manageRequirement.jsp",
-                    new ManageRequirementBean(getViewContext(), _specimenRequest, requirement), errors);
-        }
-
-        @Override
-        public boolean handlePost(final ManageRequirementForm form, BindException errors) throws Exception
-        {
-            if (!getContainer().hasPermission(getUser(), ManageRequestsPermission.class))
-                throw new UnauthorizedException("You do not have permission to update requirements!");
-
-            _specimenRequest = SpecimenRequestManager.get().getRequest(getContainer(), form.getId());
-            final SpecimenRequestRequirement requirement =
-                    SpecimenRequestRequirementProvider.get().getRequirement(getContainer(), form.getRequirementId());
-            if (_specimenRequest == null || requirement == null || requirement.getRequestId() != form.getId())
-                throw new NotFoundException();
-
-            List<AttachmentFile> files = getAttachmentFileList();
-            RequestEventType eventType;
-            StringBuilder comment = new StringBuilder();
-            comment.append(requirement.getRequirementSummary());
-            String eventSummary;
-            if (form.isComplete() != requirement.isComplete())
-            {
-                SpecimenRequestRequirement clone = requirement.createMutable();
-                clone.setComplete(form.isComplete());
-                SpecimenRequestManager.get().updateRequestRequirement(getUser(), clone);
-                eventType = RequestEventType.REQUEST_STATUS_CHANGED;
-                comment.append("\nStatus changed to ").append(form.isComplete() ? "complete" : "incomplete");
-                eventSummary = comment.toString();
-            }
-            else
-            {
-                eventType = RequestEventType.COMMENT_ADDED;
-                eventSummary = "Comment added.";
-            }
-
-            if (form.getComment() != null && form.getComment().length() > 0)
-                comment.append("\n").append(form.getComment());
-
-            SpecimenRequestEvent event;
-            try
-            {
-                event = SpecimenRequestManager.get().createRequestEvent(getUser(), requirement, eventType, comment.toString(), files);
-            }
-            catch (Exception e)
-            {
-                errors.reject(ERROR_MSG, "The request could not be updated because of an unexpected error. " +
-                        "Please report this problem to an administrator. Error details: "  + e.getMessage());
-                return false;
-            }
-            try
-            {
-
-                List<? extends NotificationRecipientSet> recipients = getUtils().getNotifications(_specimenRequest, form.getNotificationIdPairs());
-                DefaultRequestNotification notification = new DefaultRequestNotification(_specimenRequest, recipients,
-                        eventSummary, event, form.getComment(), requirement, getViewContext());
-                getUtils().sendNotification(notification, form.isEmailInactiveUsers(), errors);
-            }
-            catch (ConfigurationException | IOException e)
-            {
-                errors.reject(ERROR_MSG, "The request was updated successfully, but the notification failed: " +  e.getMessage());
-                return false;
-            }
-
-            return !errors.hasErrors();
-        }
-
-        @Override
-        public ActionURL getSuccessURL(ManageRequirementForm manageRequirementForm)
-        {
-            return getManageRequestURL(_specimenRequest.getRowId());
-        }
-
-        @Override
-        public void addNavTrail(NavTree root)
-        {
-            addSpecimenRequestNavTrail(root, _specimenRequest.getRowId());
-            root.addChild("Manage Requirement");
         }
     }
 
@@ -1232,110 +971,6 @@ public class SpecimenController extends BaseStudyController
         public void setQualityControlFlag(Boolean qualityControlFlag)
         {
             _qualityControlFlag = qualityControlFlag;
-        }
-    }
-
-    @RequiresPermission(ManageNewRequestFormPermission.class)
-    public class HandleUpdateRequestInputsAction extends FormHandlerAction<ManageRequestInputsForm>
-    {
-        @Override
-        public void validateCommand(ManageRequestInputsForm target, Errors errors)
-        {
-        }
-
-        @Override
-        public boolean handlePost(ManageRequestInputsForm form, BindException errors) throws Exception
-        {
-            SpecimenRequestInput[] inputs = new SpecimenRequestInput[form.getTitle().length];
-            for (int i = 0; i < form.getTitle().length; i++)
-            {
-                String title = form.getTitle()[i];
-                String helpText = form.getHelpText()[i];
-                inputs[i] = new SpecimenRequestInput(title, helpText, i);
-            }
-
-            if (form.getMultiline() != null)
-            {
-                for (int index : form.getMultiline())
-                    inputs[index].setMultiLine(true);
-            }
-            if (form.getRequired() != null)
-            {
-                for (int index : form.getRequired())
-                    inputs[index].setRequired(true);
-            }
-            if (form.getRememberSiteValue() != null)
-            {
-                for (int index : form.getRememberSiteValue())
-                    inputs[index].setRememberSiteValue(true);
-            }
-            SpecimenRequestManager.get().saveNewSpecimenRequestInputs(getContainer(), inputs);
-            return true;
-        }
-
-        @Override
-        public ActionURL getSuccessURL(ManageRequestInputsForm manageRequestInputsForm)
-        {
-            return getManageStudyURL();
-        }
-    }
-
-    public static final class ManageRequestInputsForm
-    {
-        private String[] _title;
-        private String[] _helpText;
-        private int[] _multiline;
-        private int[] _required;
-        private int[] _rememberSiteValue;
-
-        public String[] getHelpText()
-        {
-            return _helpText;
-        }
-
-        public void setHelpText(String[] helpText)
-        {
-            _helpText = helpText;
-        }
-
-        public String[] getTitle()
-        {
-            return _title;
-        }
-
-        public void setTitle(String[] title)
-        {
-            _title = title;
-        }
-
-        public int[] getMultiline()
-        {
-            return _multiline;
-        }
-
-        public void setMultiline(int[] multiline)
-        {
-            _multiline = multiline;
-        }
-
-        public int[] getRememberSiteValue()
-        {
-            return _rememberSiteValue;
-        }
-
-        public void setRememberSiteValue(int[] rememberSiteValue)
-        {
-            _rememberSiteValue = rememberSiteValue;
-        }
-
-        public int[] getRequired()
-        {
-            return _required;
-        }
-
-        public void setRequired(int[] required)
-        {
-            _required = required;
         }
     }
